backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
alembic
psycopg[binary]
pgvector
celery
redis
httpx
beautifulsoup4
selectolax
playwright
praw
asyncpraw
transformers
torch
sentence-transformers
scikit-learn
numpy
pandas
datasets
langdetect
langgraph
langchain
langchain-openai
slowapi
prometheus-fastapi-instrumentator
sentry-sdk
langfuse
structlog
pytest
